Add the ability to use multiple templates & paths
HAProxy allows for multiple configuration files on startup that are munged together. This commit changes consul-haproxy internally to maintain slices of templates and paths. Upon startup, it will validate the quantity of templates and paths match.

It also introduces a backwards incomaptible change to the configuration file format, changing the template and path attributes to templates and paths. These values should be a list of values instead of a single value.